Updated logic to fix CI.
